% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguinr-package.R
\docType{package}
\name{PenguinR}
\alias{PenguinR}
\alias{PenguinR-package}
\title{PenguinR: A Comprehensive Collection of Penguin Datasets for Statistical Analysis and Experimental Design}
\description{
{
Offers a comprehensive collection of penguin-related datasets suitable for descriptive statistics, hypothesis testing, and experimental design.
}
}
\details{
PenguinR: A Comprehensive Collection of University and Engineering Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of University and Engineering Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/jpablovargas340/PenguinR}
}
}
}
\author{
{
\strong{Maintainer}: Juan Pablo Vargas Perez \email{j.pablovargas340@gmail.com}
}
}
